backend/requirements.txt
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
joblib==1.0.1
Flask==2.0.1
Flask-CORS==3.0.10
seaborseaborn==0.11.2